ai-core/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
pydantic-settings==2.2.1
langchain==0.1.13
langchain-openai==0.1.1
llama-index==0.10.20
python-dotenv==1.0.1
